fix: track spawned agent shimmer focus
