+++ src/components/desktop/AgentPanelExtraAgents.tsx
@@ -30,15 +30,22 @@
  * repo cards via useAgentPanelState's runtime inventory feed.
  */
 
-import { memo, useCallback, useEffect, useMemo, useRef, useState, type MouseEvent as ReactMouseEvent } from 'react';
+import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { Folder as IconoirFolder } from 'iconoir-react';
 import { ChevronDown, ChevronRight } from '@/components/desktop/lucide-shims';
-import { AgentStatusDot, type AgentDotState } from '@/components/desktop/AgentStatusDot';
+import {
+  ExtraAgentActionMenu,
+  ExtraAgentRowView,
+  type AgentOrigin,
+  type ExtraAgentActionMenuState,
+  type ExtraAgentRow,
+  type LaneStatus,
+  type VisualStatus,
+} from './AgentPanelExtraAgentRow';
 import { callRetryPacket } from '@/lib/orchestrator/packet-actions';
 
 // ── Types ──
 
-type LaneStatus = 'idle' | 'launching' | 'running' | 'paused' | 'awaiting_input' | 'awaiting_human' | 'awaiting_orchestrator' | 'recovering' | 'reviewing' | 'merging' | 'failed' | 'completed' | 'archived';
 type LaneRuntime = 'codex' | 'claude-code' | 'gemini' | 'opencode';
 type LaneOwnership = 'managed' | 'attached';
 
@@ -77,24 +84,6 @@ interface RegisteredRepo {
   exists?: boolean;
 }
 
-type AgentOrigin = 'CLI' | 'MCP' | 'Mobile' | 'Webhook' | 'Cloud';
-type VisualStatus = 'running' | 'waiting' | 'idle' | 'error' | 'archived';
-
-interface ExtraAgentRow {
-  key: string;
-  origin: AgentOrigin;
-  status: VisualStatus;
-  runtime: string;
-  name: string;
-  subtitle: string;
-  lastActivityAt: number;
-  sessionKey: string | null;
-  repoPath: string | null;
-  packetId: string | null;
-  laneStatus: LaneStatus | null;
-  lastEventLabel: string | null;
-}
-
 interface ExtraAgentGroup {
   key: string;
   label: string;
@@ -104,18 +93,11 @@ interface ExtraAgentGroup {
 }
 
 export interface AgentPanelExtraAgentsProps {
+  activeSessionKey?: string | null;
   onSelectSession?: (sessionKey: string) => void;
 }
 
 // ── Constants ──
-
-const ORIGIN_LABELS: Record<AgentOrigin, string> = {
-  CLI: 'CLI',
-  MCP: 'MCP',
-  Mobile: 'Mobile',
-  Webhook: 'Webhook',
-  Cloud: 'Cloud',
-};
 
 const FONT = 'var(--font-sans-system)';
 
@@ -272,208 +254,6 @@ function groupRows(
   return groups;
 }
 
-// ── Small children ──
-
-function OriginChip({ origin }: { origin: AgentOrigin }) {
-  // CLI is the implicit default — unchipped. Only non-CLI origins get a chip.
-  if (origin === 'CLI') return null;
-  return (
-    <span
-      style={{
-        display: 'inline-flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        flexShrink: 0,
-        paddingTop: 1,
-        paddingBottom: 1,
-        paddingLeft: 5,
-        paddingRight: 5,
-        borderRadius: 4,
-        border: '1px solid var(--t-border-subtle)',
-        fontSize: 9,
-        fontWeight: 600,
-        textTransform: 'uppercase',
-        letterSpacing: '0.06em',
-        color: 'var(--t-text-muted)',
-        fontFamily: FONT,
-      }}
-    >
-      {ORIGIN_LABELS[origin]}
-    </span>
-  );
-}
-
-function rowDotState(row: ExtraAgentRow): AgentDotState {
-  const lane = row.laneStatus;
-  if (lane === 'completed') return 'merged';
-  if (lane === 'failed' || lane === 'recovering') return 'failed';
-  if (lane === 'archived') return 'idle';
-  if (row.status === 'running') return 'running';
-  if (row.status === 'waiting') return 'review';
-  return row.status === 'error' ? 'failed' : 'idle';
-}
-
-function rowStatusLabel(row: ExtraAgentRow): string {
-  const lane = row.laneStatus;
-  if (lane === 'reviewing') return row.lastEventLabel === 'pr_created' ? 'PR open' : 'review ready';
-  if (lane === 'awaiting_input') return 'needs input';
-  if (lane === 'awaiting_human') return 'needs you';
-  if (lane === 'awaiting_orchestrator') return 'escalated';
-  if (lane === 'failed' || lane === 'recovering') return 'failed';
-  if (lane === 'archived') return 'archived';
-  if (lane === 'completed') return 'merged';
-  if (lane === 'launching' || lane === 'running' || lane === 'merging') return 'running';
-  if (lane === 'paused' || lane === 'idle') return 'idle';
-  if (row.status === 'running') return 'running';
-  if (row.status === 'waiting') return 'review ready';
-  if (row.status === 'error') return 'failed';
-  return row.status === 'archived' ? 'archived' : 'idle';
-}
-
-function ExtraAgentRowView({
-  row,
-  onSelectSession,
-  onOpenMenu,
-  busy,
-  onRetryPacket,
-}: {
-  row: ExtraAgentRow;
-  onSelectSession?: (sessionKey: string) => void;
-  onOpenMenu?: (event: ReactMouseEvent, row: ExtraAgentRow) => void;
-  busy: boolean;
-  onRetryPacket?: (row: ExtraAgentRow) => void;
-}) {
-  // Same status vocabulary as the chat rows (AgentStatusDot): accent pulse
-  // while running — flips to the binary orbit once long-running — review
-  // orange, failed red, idle ring. Keeps the panel unified with the history list.
-  const dotState = rowDotState(row);
-  const dotLabel = rowStatusLabel(row);
-  const canFocus = Boolean(row.sessionKey && onSelectSession);
-  const canRetry = Boolean(row.packetId && onRetryPacket && (row.laneStatus === 'failed' || row.laneStatus === 'recovering'));
-  const canInteract = canFocus || canRetry;
-  const [hovered, setHovered] = useState(false);
-  const handleClick = useCallback(() => {
-    if (row.sessionKey) onSelectSession?.(row.sessionKey);
-  }, [row.sessionKey, onSelectSession]);
-
-  return (
-    <button
-      type="button"
-      disabled={!canInteract}
-      onClick={handleClick}
-      onContextMenu={(event) => {
-        event.preventDefault();
-        onOpenMenu?.(event, row);
-      }}
-      title={canFocus ? `Focus ${row.name}` : row.name}
-      style={{
-        width: '100%',
-        minHeight: 28,
-        paddingTop: 5,
-        paddingBottom: 5,
-        // Aligned with HistoryChatRow chat-text X (37) so spawned agents
-        // share the same left rail as project chats above. paddingRight
-        // matches HistoryChatRow (12) so trailing motion rings sit on
-        // the same vertical column.
-        paddingLeft: 37,
-        paddingRight: 12,
-        borderWidth: 0,
-        background: 'transparent',
-        display: 'flex',
-        alignItems: 'center',
-        gap: 9,
-        textAlign: 'left',
-        cursor: canInteract ? 'pointer' : 'default',
-        opacity: row.status === 'archived' ? 0.58 : 1,
-        fontFamily: FONT,
-      }}
-      onMouseEnter={(e) => {
-        setHovered(true);
-        if (canInteract) e.currentTarget.style.background = 'var(--t-panel-hover)';
-      }}
-      onMouseLeave={(e) => {
-        setHovered(false);
-        e.currentTarget.style.background = 'transparent';
-      }}
-    >
-      <span
-        style={{
-          flex: 1,
-          minWidth: 0,
-          fontSize: 13.5,
-          fontWeight: 300,
-          color: 'var(--t-text)',
-          letterSpacing: '-0.1px',
-          lineHeight: 1.25,
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-          whiteSpace: 'nowrap',
-        }}
-      >
-        {row.name}
-        {row.subtitle ? (
-          <span
-            style={{
-              marginLeft: 6,
-              fontSize: 9.5,
-              color: 'var(--t-text-muted)',
-              fontWeight: 260,
-              letterSpacing: '-0.4px',
-            }}
-          >
-            {row.subtitle}
-          </span>
-        ) : null}
-      </span>
-      <OriginChip origin={row.origin} />
-      {canRetry ? (
-        <span
-          title="Retry failed packet"
-          onMouseDown={(event) => event.stopPropagation()}
-          onClick={(event) => {
-            event.stopPropagation();
-            if (busy) return;
-            onRetryPacket?.(row);
-          }}
-          style={{
-            minHeight: 22,
-            borderRadius: 7,
-            paddingTop: 0,
-            paddingRight: 7,
-            paddingBottom: 0,
-            paddingLeft: 7,
-            display: 'inline-flex',
-            alignItems: 'center',
-            background: 'transparent',
-            color: busy ? 'var(--t-text-faint)' : 'var(--t-text-muted)',
-            cursor: busy ? 'default' : 'pointer',
-            fontFamily: FONT,
-            fontSize: 10.5,
-            fontWeight: 300,
-            letterSpacing: '-0.1px',
-            opacity: hovered ? 1 : 0,
-            pointerEvents: hovered ? 'auto' : 'none',
-            transition: 'opacity 120ms ease, background 120ms ease, color 120ms ease',
-            flexShrink: 0,
-          }}
-          onMouseEnter={(event) => {
-            if (busy) return;
-            event.currentTarget.style.background = 'var(--t-hover)';
-            event.currentTarget.style.color = 'var(--t-text)';
-          }}
-          onMouseLeave={(event) => {
-            event.currentTarget.style.background = 'transparent';
-            event.currentTarget.style.color = busy ? 'var(--t-text-faint)' : 'var(--t-text-muted)';
-          }}
-        >
-          {busy ? 'Retrying' : 'Retry'}
-        </span>
-      ) : null}
-      <AgentStatusDot state={dotState} label={dotLabel} />
-    </button>
-  );
-}
-
 function GroupHeader({
   label,
   count,
@@ -595,17 +375,11 @@ function GroupHeader({
 
 const COLLAPSED_KEY = 'o8:agent-panel:spawned-agents-collapsed';
 
-interface ActionMenuState {
-  x: number;
-  y: number;
-  row: ExtraAgentRow;
-}
-
-function AgentPanelExtraAgentsBase({ onSelectSession }: AgentPanelExtraAgentsProps) {
+function AgentPanelExtraAgentsBase({ activeSessionKey, onSelectSession }: AgentPanelExtraAgentsProps) {
   const [lanes, setLanes] = useState<LaneSummary[]>([]);
   const [agents, setAgents] = useState<AgentSummary[]>([]);
   const [repos, setRepos] = useState<RegisteredRepo[]>([]);
-  const [actionMenu, setActionMenu] = useState<ActionMenuState | null>(null);
+  const [actionMenu, setActionMenu] = useState<ExtraAgentActionMenuState | null>(null);
   const [archivedSessionKeys, setArchivedSessionKeys] = useState<Set<string>>(() => new Set());
   const [busy, setBusy] = useState(false);
   const [collapsed, setCollapsed] = useState<boolean>(() => {
@@ -750,6 +524,7 @@ function AgentPanelExtraAgentsBase({ onSelectSession }: AgentPanelExtraAgentsPro
             <ExtraAgentRowView
               key={row.key}
               row={row}
+              active={Boolean(activeSessionKey && row.sessionKey === activeSessionKey)}
               busy={busy}
               onSelectSession={onSelectSession}
               onRetryPacket={handleRetry}
@@ -777,141 +552,6 @@ function AgentPanelExtraAgentsBase({ onSelectSession }: AgentPanelExtraAgentsPro
         />
       ) : null}
     </section>
-  );
-}
-
-function ExtraAgentActionMenu({
-  state,
-  busy,
-  canFocus,
-  onClose,
-  onFocus,
-  onArchive,
-}: {
-  state: ActionMenuState;
-  busy: boolean;
-  canFocus: boolean;
-  onClose: () => void;
-  onFocus: () => void;
-  onArchive: () => void;
-}) {
-  useEffect(() => {
-    const onKey = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') onClose();
-    };
-    window.addEventListener('keydown', onKey);
-    return () => window.removeEventListener('keydown', onKey);
-  }, [onClose]);
-
-  const menuWidth = 190;
-  const menuHeight = 130;
-  const panelRect = typeof document === 'undefined'
-    ? null
-    : document.querySelector('[data-o8-agent-panel="true"]')?.getBoundingClientRect() ?? null;
-  const viewportWidth = typeof window === 'undefined' ? 1200 : window.innerWidth;
-  const viewportHeight = typeof window === 'undefined' ? 800 : window.innerHeight;
-  const boundaryLeft = panelRect?.left ?? 0;
-  const boundaryRight = panelRect?.right ?? viewportWidth;
-  const boundaryTop = panelRect?.top ?? 0;
-  const boundaryBottom = panelRect?.bottom ?? viewportHeight;
-  const minLeft = boundaryLeft + 8;
-  const maxLeft = Math.max(minLeft, boundaryRight - menuWidth - 8);
-  const left = Math.min(Math.max(state.x, minLeft), maxLeft);
-  const minTop = boundaryTop + 8;
-  const maxTop = Math.max(minTop, boundaryBottom - menuHeight - 8);
-  const top = Math.min(Math.max(state.y, minTop), maxTop);
-
-  return (
-    <>
-      <button
-        type="button"
-        aria-label="Close spawned agent action menu"
-        onClick={onClose}
-        style={{ position: 'fixed', inset: 0, zIndex: 58, border: 0, background: 'transparent', cursor: 'default' }}
-      />
-      <div
-        style={{
-          position: 'fixed',
-          left,
-          top,
-          zIndex: 59,
-          width: menuWidth,
-          borderRadius: 13,
-          border: '1px solid var(--t-divider-subtle)',
-          background: 'color-mix(in srgb, var(--t-bg-elevated, #ffffff) 86%, transparent)',
-          boxShadow: '0 18px 48px rgba(15, 23, 42, 0.12)',
-          backdropFilter: 'blur(18px) saturate(145%)',
-          WebkitBackdropFilter: 'blur(18px) saturate(145%)',
-          padding: 7,
-          color: 'var(--t-text)',
-        }}
-      >
-        <div style={{ padding: '4px 7px 7px' }}>
-          <div style={{ fontSize: 11.25, lineHeight: '15px', fontWeight: 620, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
-            {state.row.name}
-          </div>
-          <div style={{ marginTop: 1, color: 'var(--t-text-faint)', fontSize: 10, lineHeight: '13px', overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
-            {state.row.subtitle || state.row.runtime}
-          </div>
-        </div>
-        <div style={{ display: 'grid', gap: 2 }}>
-          <ExtraAgentMenuRow label="Open" disabled={!canFocus || busy} onClick={onFocus} />
-          <ExtraAgentMenuRow label="Archive" disabled={busy || !state.row.sessionKey} onClick={onArchive} />
-        </div>
-      </div>
-    </>
-  );
-}
-
-function ExtraAgentMenuRow({
-  label,
-  disabled = false,
-  danger = false,
-  onClick,
-}: {
-  label: string;
-  disabled?: boolean;
-  danger?: boolean;
-  onClick: () => void;
-}) {
-  return (
-    <button
-      type="button"
-      disabled={disabled}
-      onClick={(event) => {
-        event.stopPropagation();
-        onClick();
-      }}
-      style={{
-        width: '100%',
-        minHeight: 29,
-        borderRadius: 9,
-        border: 0,
-        background: 'transparent',
-        color: disabled ? 'var(--t-text-faint)' : danger ? '#dc2626' : 'var(--t-text-muted)',
-        cursor: disabled ? 'default' : 'pointer',
-        textAlign: 'left',
-        paddingTop: 0,
-        paddingRight: 9,
-        paddingBottom: 0,
-        paddingLeft: 9,
-        fontSize: 11.25,
-        lineHeight: '15px',
-        fontWeight: danger ? 620 : 560,
-        transition: 'background 120ms ease, color 120ms ease',
-      }}
-      onMouseEnter={(event) => {
-        if (disabled) return;
-        event.currentTarget.style.background = 'var(--t-hover)';
-        event.currentTarget.style.color = danger ? '#dc2626' : 'var(--t-text)';
-      }}
-      onMouseLeave={(event) => {
-        event.currentTarget.style.background = 'transparent';
-        event.currentTarget.style.color = disabled ? 'var(--t-text-faint)' : danger ? '#dc2626' : 'var(--t-text-muted)';
-      }}
-    >
-      {label}
-    </button>
   );
 }
